# openapi.json
{"openapi": "3.1.0", "info": {"title": "Resume Generator API", "description": "Generate and evaluate professional resumes by job category.", "version": "1.0.0"}, "paths": {"/health": {"get": {"tags": ["System"], "summary": "Health Check", "description": "Backend health check \u2014 confirms the API is running.", "operationId": "health_check_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HealthResponse"}}}}}}}, "/categories": {"get": {"tags": ["Resume"], "summary": "Get Categories", "description": "Return all available resume categories discovered from the data directory.\nEach category corresponds to a folder in data/data/.", "operationId": "get_categories_categories_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategoryListResponse"}}}}}}}, "/ingest": {"post": {"tags": ["System"], "summary": "Run Ingestion", "description": "Trigger the full ingestion pipeline:\nLoad PDFs \u2192 Clean PII \u2192 Chunk \u2192 Embed \u2192 Store in ChromaDB.\n\nThis runs as a background task and returns immediately.\nFor large datasets, check server logs for progress.", "operationId": "run_ingestion_ingest_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IngestResponse"}}}}}}}, "/generate-resume": {"post": {"tags": ["Resume"], "summary": "Generate Resume Endpoint", "description": "Generate a professional anonymized resume for the given category.\n\nPipeline:\n  1. Retrieve relevant chunks from ChromaDB (filtered by category)\n  2. Apply MMR retrieval + CrossEncoder reranking\n  3. Call NVIDIA NIM LLM to generate resume text\n  4. Convert to PDF and save\n  5. Return resume text + PDF download URL\n\nRequest body:\n  { \"category\": \"INFORMATION-TECHNOLOGY\" }", "operationId": "generate_resume_endpoint_generate_resume_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateResumeRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateResumeResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/download/{filename}": {"get": {"tags": ["Resume"], "summary": "Download Pdf", "description": "Download a previously generated resume PDF.\nFiles are stored in backend/generated_resumes/.", "operationId": "download_pdf_download__filename__get", "parameters": [{"name": "filename", "in": "path", "required": true, "schema": {"type": "string", "title": "Filename"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/evaluate/all": {"get": {"tags": ["Evaluation"], "summary": "Evaluate All", "description": "Evaluate accuracy for all categories (served from cache when available).", "operationId": "evaluate_all_evaluate_all_get", "parameters": [{"name": "regenerate", "in": "query", "required": false, "schema": {"type": "boolean", "description": "Regenerate resumes for every category before evaluation", "default": false, "title": "Regenerate"}, "description": "Regenerate resumes for every category before evaluation"}, {"name": "force_refresh", "in": "query", "required": false, "schema": {"type": "boolean", "description": "Ignore cache and recompute all 24 category scores", "default": false, "title": "Force Refresh"}, "description": "Ignore cache and recompute all 24 category scores"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluateAllResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/evaluate/{category}": {"get": {"tags": ["Evaluation"], "summary": "Evaluate Single Category", "description": "Evaluate accuracy for one resume category.", "operationId": "evaluate_single_category_evaluate__category__get", "parameters": [{"name": "category", "in": "path", "required": true, "schema": {"type": "string", "title": "Category"}}, {"name": "regenerate", "in": "query", "required": false, "schema": {"type": "boolean", "description": "Force new LLM resume generation instead of using cached PDF", "default": false, "title": "Regenerate"}, "description": "Force new LLM resume generation instead of using cached PDF"}, {"name": "force_refresh", "in": "query", "required": false, "schema": {"type": "boolean", "description": "Ignore cache and recompute accuracy for this category", "default": false, "title": "Force Refresh"}, "description": "Ignore cache and recompute accuracy for this category"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccuracyReportResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/accuracy/report": {"get": {"tags": ["Evaluation"], "summary": "Get Accuracy Report", "description": "Return the latest accuracy report (single category or aggregate summary).", "operationId": "get_accuracy_report_accuracy_report_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccuracyReportResponse"}}}}}}}, "/accuracy/report/all": {"get": {"tags": ["Evaluation"], "summary": "Get All Accuracy Report", "description": "Return cached evaluate-all report (instant; no re-evaluation).", "operationId": "get_all_accuracy_report_accuracy_report_all_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EvaluateAllResponse"}}}}}}}, "/accuracy/cache/status": {"get": {"tags": ["Evaluation"], "summary": "Get Accuracy Cache Status", "description": "Whether a cached all-categories accuracy report is available.", "operationId": "get_accuracy_cache_status_accuracy_cache_status_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/accuracy/cache/{category}": {"get": {"tags": ["Evaluation"], "summary": "Get Category Cached Report", "description": "Return cached accuracy for one category without re-running evaluation.", "operationId": "get_category_cached_report_accuracy_cache__category__get", "parameters": [{"name": "category", "in": "path", "required": true, "schema": {"type": "string", "title": "Category"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AccuracyReportResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"AccuracyReportResponse": {"properties": {"category": {"type": "string", "title": "Category"}, "retrieval_accuracy": {"type": "string", "title": "Retrieval Accuracy"}, "category_accuracy": {"type": "string", "title": "Category Accuracy"}, "content_accuracy": {"type": "string", "title": "Content Accuracy"}, "privacy_accuracy": {"type": "string", "title": "Privacy Accuracy"}, "overall_accuracy": {"type": "string", "title": "Overall Accuracy"}, "evaluated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Evaluated At"}, "error": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Error"}}, "type": "object", "required": ["category", "retrieval_accuracy", "category_accuracy", "content_accuracy", "privacy_accuracy", "overall_accuracy"], "title": "AccuracyReportResponse", "description": "Accuracy evaluation report for one category."}, "CategoryListResponse": {"properties": {"categories": {"items": {"type": "string"}, "type": "array", "title": "Categories"}, "total": {"type": "integer", "title": "Total"}}, "type": "object", "required": ["categories", "total"], "title": "CategoryListResponse", "description": "Response body for GET /categories."}, "EvaluateAllResponse": {"properties": {"categories": {"items": {"additionalProperties": true, "type": "object"}, "type": "array", "title": "Categories"}, "summary": {"$ref": "#/components/schemas/EvaluateAllSummary"}, "evaluated_at": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Evaluated At"}, "total_categories": {"type": "integer", "title": "Total Categories"}}, "type": "object", "required": ["categories", "summary", "total_categories"], "title": "EvaluateAllResponse", "description": "Accuracy evaluation across all categories."}, "EvaluateAllSummary": {"properties": {"retrieval_accuracy": {"type": "string", "title": "Retrieval Accuracy"}, "category_accuracy": {"type": "string", "title": "Category Accuracy"}, "content_accuracy": {"type": "string", "title": "Content Accuracy"}, "privacy_accuracy": {"type": "string", "title": "Privacy Accuracy"}, "overall_accuracy": {"type": "string", "title": "Overall Accuracy"}}, "type": "object", "required": ["retrieval_accuracy", "category_accuracy", "content_accuracy", "privacy_accuracy", "overall_accuracy"], "title": "EvaluateAllSummary"}, "GenerateResumeRequest": {"properties": {"category": {"type": "string", "title": "Category", "description": "Resume category name (e.g., 'INFORMATION-TECHNOLOGY')", "example": "INFORMATION-TECHNOLOGY"}}, "type": "object", "required": ["category"], "title": "GenerateResumeRequest", "description": "Request body for POST /generate-resume."}, "GenerateResumeResponse": {"properties": {"category": {"type": "string", "title": "Category"}, "resume_text": {"type": "string", "title": "Resume Text"}, "pdf_url": {"type": "string", "title": "Pdf Url"}}, "type": "object", "required": ["category", "resume_text", "pdf_url"], "title": "GenerateResumeResponse", "description": "Response body for POST /generate-resume."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "HealthResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "message": {"type": "string", "title": "Message"}, "version": {"type": "string", "title": "Version", "default": "1.0.0"}}, "type": "object", "required": ["status", "message"], "title": "HealthResponse", "description": "Response body for GET /health."}, "IngestResponse": {"properties": {"status": {"type": "string", "title": "Status"}, "message": {"type": "string", "title": "Message"}, "categories_processed": {"type": "integer", "title": "Categories Processed"}, "total_chunks": {"type": "integer", "title": "Total Chunks"}, "errors": {"items": {"type": "string"}, "type": "array", "title": "Errors", "default": []}}, "type": "object", "required": ["status", "message", "categories_processed", "total_chunks"], "title": "IngestResponse", "description": "Response body for POST /ingest."}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}